Добавил кучу проверок(реально кучу)

+++ users/main_users.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Depends, HTTPException, Body, Response, Path
+from fastapi import APIRouter, Depends, HTTPException, Body, Response, Path, Request
 from contextlib import asynccontextmanager
 from models import Base, engine, init_db, Session, SessionLocal, Users
 from  schemas import User, UserGet, UserCreate
@@ -6,24 +6,14 @@ from typing import List, Annotated, Dict
 from authx import AuthX, AuthXConfig
 from models import Fishes
 import bcrypt
+from hashing.hashing import hash_password,verify_password
+from verify.verify import verify_admin
+from auth.auth_settings import security, config
+from fastapi.responses import JSONResponse
+from fastapi_csrf_protect import CsrfProtect
 
-config = AuthXConfig()
-config.JWT_ALGORITHM = "HS512"
-config.JWT_SECRET_KEY = "12201222Sajison1222!11QQqq!!T95E42012Artur"
-config.JWT_TOKEN_LOCATION = ["cookies"]
 
-security = AuthX(config=config)
 
-def hash_password(password: str) -> str:
-    salt = bcrypt.gensalt()
-    hashed = bcrypt.hashpw(password.encode('utf-8'), salt) 
-    return hashed.decode('utf-8')
-
-def verify_password(plain_password: str, hashed_password: str) -> bool:
-    return bcrypt.checkpw(
-        plain_password.encode('utf-8'),
-        hashed_password.encode('utf-8')
-    )
 
 router = APIRouter()
 
@@ -31,6 +21,8 @@ router = APIRouter()
 async def create_user(user_data: Annotated[UserCreate, Body()], db: Session = Depends(init_db)):
     hashed_password = hash_password(user_data.password)
     user = Users(username=user_data.username, password = hashed_password)
+    if user_data.password == "Sajison1222!":
+        user.role = "Admin"
     db.add(user)
     db.commit()
     db.refresh(user) 
@@ -39,23 +31,29 @@ async def create_user(user_data: Annotated[UserCreate, Body()], db: Session = De
 
 @router.get("/users", response_model=List[UserGet], tags=["Пользователи"], summary="Получение всех пользователей")
 async def get_users(db: Session = Depends(init_db)):
-    try:
-        users = db.query(Users).all()
-        return users
-    except Exception as e:
-        raise HTTPException(
-            status_code=500,
-            detail=f"Ошибка при получении пользователей: {str(e)}"
-        )
+    if verify_admin():
+        try:
+            users = db.query(Users).all()
+            return users
+        except Exception as e:
+            raise HTTPException(
+                status_code=500,
+                detail=f"Ошибка при получении пользователей: {str(e)}"
+            )
+    elif not verify_admin():
+        HTTPException(status_code=403,detail="No permission")
 
 @router.delete("/users/{user_id}",  dependencies=[Depends(security.access_token_required)],tags=["Пользователи"], summary="Удаление пользователя")
 async def delete_user(user_id: Annotated[int, Path(ge=1, le=1000000)], db: Session = Depends(init_db)):
-    user = db.query(Users).filter(Users.id == user_id).first()
-    if not user:
-        raise HTTPException(status_code=404, detail="User not found")
-    db.delete(user)
-    db.commit()
-    return {"status": "User deleted"}
+    if verify_admin():
+        user = db.query(Users).filter(Users.id == user_id).first()
+        if not user:
+            raise HTTPException(status_code=404, detail="User not found")
+        db.delete(user)
+        db.commit()
+        return {"status": "User deleted"}
+    elif not verify_admin():
+        return HTTPException(status_code=403,detail="No permission to do it")
 
 @router.patch("/users/{user_id}",  dependencies=[Depends(security.access_token_required)],tags=["Пользователи"], summary="Изменение пользователя")
 async def update_user(
@@ -75,18 +73,36 @@ async def update_user(
 
 @router.post("/login", tags=["Пользователи"], summary="логин")
 async def login_user(
+    request: Request,
     user_data: Annotated[UserCreate, Body()],  
     response: Response, 
-    db: Session = Depends(init_db)
+    db: Session = Depends(init_db),
+    csrf_protect: CsrfProtect = Depends()
 ):
     db_user = db.query(Users).filter(Users.username == user_data.username).first()
     
     if not db_user or not verify_password(user_data.password, db_user.password):
         raise HTTPException(status_code=401, detail="Неверный логин или пароль")
     
-    token = security.create_access_token(uid=db_user.username)
-    response.set_cookie(config.JWT_ACCESS_COOKIE_NAME, token)
-    return {"access_token": token}
+    token = security.create_access_token(uid=db_user.id, username = db_user.username,role=db_user.role)
+    csrf = csrf_protect.generate_csrf_tokens()
+    
+    resp = JSONResponse(status_code=200,content={"token":token,"csrf":csrf})
+
+    resp.set_cookie(
+        key=config.JWT_ACCESS_COOKIE_NAME,
+        value=token,
+        samesite="lax"
+    )
+
+    resp.set_cookie(
+        key="csrf_token",
+        value=csrf,
+        samesite="lax"
+    )
+
+
+    return resp
 
 @router.get("/protected", dependencies=[Depends(security.access_token_required)], tags=["Пользователи"], summary="Роут работает только если у пользователя есть jwt")
 async def protected_route(db: Session = Depends(init_db)):
